refactor: 이벤트리스너를 이벤트위임으로 리팩토링
