+++ README.md
@@ -97,3 +97,6 @@ LottoController.js
 [ ] 14. 숫자가 아닌 입력 시 에러메세지 추가
 [ ] 15. 써밋: 2글자 이상 입력시 바로 다음으로 넘어가는 게 꼭 좋은 경험일까?
 [ ] 16. 사용자 친화적 메세지로 가공
+[ ] 17. 도큐먼트 업데이트 자제. 셀렉터도 자제.
+[ ] 18. 당첨 번호 첫 입력칸으로 바로 focus() 되게
+[ ] 19. 처음에 박스 움직이는 거 신경쓰임. error handling 다시
